container.go should support docker versions 1.8 and 1.9
docker 1.8 changed the structure of the docker inspect output, no longer
including .Volumes.  Instead it's in .Mounts with a different structure.

In docker 1.8 the command `docker inspect --format {{.Volumes}} CONTAINER`
will just return "<no value>".  But, in docker 1.9 it will be an error.

This commit checks for either the 1.8 or 1.9 response to requesting .Volumes
and will use .Mounts in those cases.  Else (docker 1.7 and earlier) .Volumes
will be used as before.